Update README.md
Description:
CommunityConnect is a comprehensive platform designed for Non-Governmental Organisations (NGOs) seeking assistance, both financial and non-financial, from socially conscious companies interested in making a meaningful impact. The platform that help to build a sustained and effective communication between the NG)s and the corporate entities. Introduction Despite the flourishing ecosystem, identifying the right organisation to support can be a challenging endeavor. Traditionally, the path to partnership is strewn with obstacles, from exhaustive due diligence to the inherent limitations in an NGO's outreach capabilities. Recognizing this, CommunityConnect offers an intuitive, data-driven interface that transcends these barriers, enabling seamless identification, communication, and alignment of missions between NGOs and corporate entities.